Ticket #217 http://opentripplanner.org/ticket/217 Upgraded ExtJS to 3.2.1 and OpenLayers to 2.9 Tested on Windows Browsers (IE 8, FF 3.6, Safari 4.0.5, Opera 9, Chrome 4.1).
NOTE: IE is the only browser that is broken - funky images (sometimes not transparency; image maps not doing what CSS tells them to do, etc...) and no south panel for the elevation graph.  This has nothing to do with the upgrade, as I had seen the same problems in the ExtJS 3.0 and OL 2.8 versions that were previously running.