Fixes #259
Adds a dialog box to each itinerary that provides links back to OTP.  Configuration allows multiple links to be identified.  This will be used for comparing trips out of OTP with Google, etc...

Also note that the configuration is currently written for Portland / TriMet...and a future configuration edit will make this more generic.