remove TurnEdge fromv and tov fields
TurnEdge has fromv and tov fields that are not of type Vertex. These
fields simply hide the fields in the abstract base class, which still
exist and waste memory. 

The quick fix here is to cast fromv to (TurnVertex) every time it is
referenced within this class. The long-term fix will involve factoring
out a StreetSegment interface that will be implemented by TurnVertex and
PlainStreetEdge.