remove conflicting (and apparently unnecessary) init-param
it seems that only one instance of each named init-param is used.
which one is used depends on the servlet container or some other detail.

JSON response generation can be enabled by adding a maven dependency on
jersey-json, and the package in the removed init-param does not
seem to provide this feature directly.