Add support for the new "bikes_allowed" field in routes.txt and trips.txt.
The trips.txt "trip_bikes_allowed" proposal has been around for a while now
and is supported by OTP.  As part of the discussion to formally adopt the proposal
into the official GTFS spec, it was proposed that we change the semantics of the
field to match the existing "wheelchair_accessible" field (1=allowed, 2=not allowed).
To avoid breaking existing feeds and support phased migration, we proposed changing
the field name to "bikes_allowed" as well.  See discussion at:

https://groups.google.com/d/msg/gtfs-changes/rEiSeKNc4cs/gTTnQ_yXtPgJ

This commit refactors bike access semantics in OTP to support both the new and old
field, abstracting bike access into a new BikeAccess enum.