Turn off zoom-to-result by default
This client has floating windows which almost always obscure parts of
the path or the endpoints. We should perhaps make the zoom behavior
conditional so it only happens when a new path falls outside the current
window frame. It might also make sense to "dock" the windows into the
border and reduce the size of the map.