check for the presence of the correct transfer stop, not the length of the path.
for some reason several of these path lengths are different than what was originally expected.
the only hints given in the test about the correct results is that they must pass
through stop C. The code has therefore been changed to check for that condition instead
of checking the path length.