remove OTPPriorityQueue interface and accompanying factory interface
use of pluggable factories makes the code less clean and readable.
we only ever use the BinHeap implementation because it represents a
good trade-off between speed and flexibility. it is also simpler,
smaller, and at least as fast as a Fibonacci heap, which gains lower
theoretical complexity at the cost of space and code complexity.

Also simplified the priority queue tests, which were being used
as a speed comparison.