A number of changes necessary to lighten the weight of analyst objects
The general problem is that Analyst objects retained references to lots of objects so they should cache things that might come in handy some time in their future. This makes it difficult to serialize the objects for storage or transmission. An alternative strategy is to ask the environment to cache handy objects while analyst objects like PointSet and SampleSet remain fairly svelte message objects.