Implements the option --preFlight
Currently --inMemory is able to load a GTFS file from disk, but explicitly won't write Graph.obj to disk. Our usecase in instance of OpenTripPlanner running in preFlight mode. This mode will compile a graph upon startup, and when
the service is accessibily will be used to evaluate the quality of the running instance. Once the running instance passes the criteria the Graph.obj will be deployed to its final location, where other instances are able to start
from it.