remove TestRequest
Most times I make a modification to OTP routing one of these tests
cryptically fails and it takes me an hour or two to figure out why.
These tests are expecting very specific results (trip IDs, numbers
of legs) in results from real-world graphs. The test fixtures are
not controlled, and there is no real documentation explaining
what's going on. It begins with a comment explaining that the
implementation is a hack.

We do need to test these things, but in well-documented, purpose
built tests, and some of them are integration tests rather than
unit tests. Most should be covered by an automated run of OTPQA.