remove site element so Maven doesn't generate Javadoac on release
release:perform checks out the last tag, then runs deploy and deploy-site
but it only runs deploy-site if a site is defined in the POM.
For some reason deploy-site was using a different version of the javadoc
plugin, not switching off the lint checking, and aborting the deployment.

We weren't even using this site (it's non-versioned) so I just removed it.