RouteTextColors are generated based on routeColors
GTFS route can have route color and route text color. This colors are
used to show route in a specific color. Both values should be set but
sometimes aren't. So we generate textColor if it doesn't exist but
routeColor does.

Generated textColor can be black or white.

If routeColor is bright it is black otherwise it is white.
Brightness is calculated based on luminance formula