Changes to SeattleFareServiceFactory so fares can be specified in GTFS.
We would like for fares in the Puget Sound region to be specified in GTFS
rather than hardcoded into SeattleFareServiceFactory.java. This will allow
fare changes to be rolled out without a complete rebuild of OTP. Changes in
how fares are calculated may still require new code, but simple changes in
price such as the KCM $0.25 fare hike on 3/1 will not.

- Remove hardcoded fares. (These are being added to GTFS upstream)
- Add senior/youth fare prices as extra column in fare attributes. See
pull request to conveyal/onebusaway-gtfs-modules.
- Remove processGTFS agency separation. No longer necessary because of new
feed ID feature in OpenTripPlanner.