Use new SimpleStreetLinker on origin/destination
Working on #2267. Linking works but code is a little messy.
Currently one temporary vertex is created when closest thing is
existing from/to vertex. And a second split vertex is created if
the closest edge is a street which is splitted. All new edges and
vertices are temporary and are removed when thread finishes.